requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
alembic==1.12.1
mysqlclient==2.2.0
minio==7.2.0
pydantic-settings==2.1.0
email-validator==2.1.0.post1
elasticsearch==8.11.0
mysql-connector-python==8.2.0
pymysql==1.1.0
aiohttp==3.9.1